requirements.txt
Flask
flask-cors
gunicorn
face-recognition
opencv-python
numpy
Pillow